test GetStateStorageBlockHeight
