+++ test/builders/statediffs.go
@@ -1,8 +1,8 @@
 package builders
 
 import (
-	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
+	"github.com/orbs-network/orbs-spec/types/go/protocol"
 )
 
 type sdiff struct {
