+++ test/harness/instrumentation/reporting.go
@@ -42,7 +42,6 @@ func (l *latch) Infof(message string, params ...interface{}) {
 	l.Info(fmt.Sprintf(message, params...))
 }
 
-
 func (l *latch) Error(err error) {
 	l.Info(err.Error())
 }
@@ -77,7 +76,6 @@ func (e *bufferedLog) Info(message string) {
 func (e *bufferedLog) Infof(message string, params ...interface{}) {
 	e.Info(fmt.Sprintf(message, params...))
 }
-
 
 func (e *bufferedLog) Error(err error) {
 	e.log(err.Error())
